fix: rename style.js -> object.js and move the function to a separate file.
